Fix invalid column dereference
If we were unable to fetch a column, we return early from the mysql
column fetch logic. However, we've already set the number of columns in
the result object. This causes a crash during the eventual
mysql-to-dbwrap row/column conversion logic. The conversion logic
(rightly) assumes that we have all available columns, when, in fact, we
don't.

As a temporary workaround (until I get time for a proper fix), simply
set the number of columns to zero (0) so that the conversion logic
doesn't attempt to reference columns that aren't there.

Signed-off-by:	Shawn Webb <swebb@blackhawknest.com>